Extract unmutes detector
